utils refactor and unit tests
