--- src/exchanges/huobi/utils/functions/other.js
@@ -1,13 +1,0 @@
-const { createHmac } = require('crypto');
-
-const createHmacSignature = (method, url, queryString, privateKey) => {
-    const preSignedText = `${method}\napi.huobi.pro\n/${url}\n${queryString.toString()}`;
-
-    return createHmac('sha256', privateKey)
-        .update(preSignedText)
-        .digest('base64');
-};
-
-module.exports = {
-    createHmacSignature,
-};
